refactor: ResourceManagerT の deps 統一・ファイル統合・インライン化
- ResourceManagerDeps を deps_ メンバ1つに統一し、個別ポインタ7つを削除
- resource_manager_impl.h を resource_manager.h に統合
- 全メンバ関数をクラス本体にインライン定義
- noexcept/constexpr を付与可能な箇所に追加
